# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Employee CRUD API","version":"0.1.0"},"paths":{"/employees":{"get":{"summary":"Read Employees","operationId":"read_employees_employees_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Employee"},"type":"array","title":"Response Read Employees Employees Get"}}}}}},"post":{"summary":"Create Employee","operationId":"create_employee_employees_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/employees/{employee_id}":{"get":{"summary":"Read Employee","operationId":"read_employee_employees__employee_id__get","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","title":"Employee Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Employee","operationId":"update_employee_employees__employee_id__put","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","title":"Employee Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Employee","operationId":"delete_employee_employees__employee_id__delete","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","title":"Employee Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Employee":{"properties":{"name":{"type":"string","title":"Name"},"age":{"type":"integer","title":"Age"},"city":{"type":"string","title":"City"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","age","city","id"],"title":"Employee"},"EmployeeCreate":{"properties":{"name":{"type":"string","title":"Name"},"age":{"type":"integer","title":"Age"},"city":{"type":"string","title":"City"}},"type":"object","required":["name","age","city"],"title":"EmployeeCreate"},"EmployeeUpdate":{"properties":{"name":{"type":"string","title":"Name"},"age":{"type":"integer","title":"Age"},"city":{"type":"string","title":"City"}},"type":"object","required":["name","age","city"],"title":"EmployeeUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}